+++ .babelrc
@@ -1,6 +1,6 @@
 {
   "presets": [
-    ["es2015", {"modules": false}],
+    ["env", {"modules": false}],
     "react",
     "stage-2"
   ],
