+++ package.json
@@ -20,12 +20,12 @@
   "author": "Naoki Sawada",
   "license": "MIT",
   "dependencies": {
-    "moducks": "^0.2.3",
+    "moducks": "^1.0.0-beta.3",
     "react": "^16.2.0",
     "react-css-modules": "^4.1.0",
     "react-dom": "^16.2.0",
-    "react-redux": "^5.0.1",
-    "redux": "^3.6.0",
+    "react-redux": "^5.0.7",
+    "redux": "^4.0.0",
     "redux-actions": "^2.0.1",
     "redux-logger": "^3.0.6",
     "redux-saga": "^0.16.0"
@@ -36,7 +36,7 @@
     "babel-plugin-root-import": "^5.1.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.16.0",
     "babel-preset-stage-2": "^6.24.1",
     "browser-sync": "^2.18.6",
@@ -46,7 +46,7 @@
     "style-loader": "^0.20.2",
     "uglifyjs-webpack-plugin": "^1.2.2",
     "url-loader": "^1.0.1",
-    "webpack": "^4.1.0",
-    "webpack-cli": "^2.0.10"
+    "webpack": "^4.12.0",
+    "webpack-cli": "^3.0.3"
   }
 }
