+++ src/js/app.js
@@ -1,8 +1,8 @@
 import { Provider } from 'react-redux';
-import createStore from '~/store';
+import configureStore from './store';
 import Main from './components/Main';
 
-const store = createStore({});
+const store = configureStore({});
 
 ReactDOM.render(
   <Provider store={store}>
