+++ src/js/modules/moducks.js
@@ -1,0 +1,4 @@
+import Moducks from 'moducks';
+import * as effects from 'redux-saga/effects';
+
+export default new Moducks({ effects: { ...effects } });
